+++ projects/rocprofiler-systems/source/lib/core/config.hpp
@@ -187,6 +187,9 @@ get_use_causal() ROCPROFSYS_HOT;
 bool
 get_use_amd_smi() ROCPROFSYS_HOT;
 
+bool
+get_use_hipfile() ROCPROFSYS_HOT;
+
 bool&
 get_use_sampling() ROCPROFSYS_HOT;