+++ projects/rocprofiler-systems/source/lib/rocprof-sys-user/rocprofiler-systems/categories.h
@@ -127,6 +127,7 @@ extern "C"
         ROCPROFSYS_CATEGORY_OVERFLOW_SAMPLING,
         ROCPROFSYS_CATEGORY_UNIFIED_MEMORY_MIGRATION_THROUGHPUT,
         ROCPROFSYS_CATEGORY_UNIFIED_MEMORY_FAULT_RATE,
+        ROCPROFSYS_CATEGORY_HIPFILE,
         ROCPROFSYS_CATEGORY_LAST
         // the value of below enum is used for iterating
         // over the enum in C++ templates. It MUST